+++ src/snowplowderby/Player.cpp
@@ -2,14 +2,14 @@
 
 using namespace snowplowderby;
 
-Player::Player(long id, b2Body* body) : id(id), body(body), user_data{USERDATA_TYPE_PLAYER, this} {
+Player::Player(short id, b2Body* body) : id(id), body(body), user_data{USERDATA_TYPE_PLAYER, this} {
 }
 
 util::UserDataWrapper* Player::get_user_data() {
     return &user_data;
 }
 
-long Player::get_id() {
+short Player::get_id() {
     return id;
 }
